Move all scripts to a proper module
