--- pretty/__init__.py
@@ -1,3 +1,0 @@
-from .collection import main as pretty_collection
-from .history import main as pretty_history
-from .playlists import main as pretty_playlists
