+++ pretty_collection.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from stemgen.ableton import main
+from stemgen.pretty_collection import main
 
 if __name__ == "__main__":
     main()
