+++ pretty_history.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from stemgen.ableton import main
+from stemgen.pretty_history import main
 
 if __name__ == "__main__":
     main()
