+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = stemgen
-version = 1.0.0
+version = 2.0.0
 author = Axel Delafosse
 author_email = axeldelafosse@gmail.com
 description = Stemgen is a Stem file generator. Convert any track into a stem and have fun with Traktor.
@@ -20,18 +20,26 @@ packages = find:
 python_requires = >=3.9
 install_requires =
     mutagen
+    demucs
+    torch
+    soundfile
+    pyloudnorm
+    traktor-nml-utils
+    pyautogui
+    pylive
+    opencv-python
 
 [options.package_data]
 * = *.json
 
 [options.entry_points]
 console_scripts =
-    stemgen = cli.stemgen:main
-    stem = cli.stem:main
-    stemsep = cli.stemsep:main
-    stemtag = cli.stemtag:main
-    stemcopy = cli.stemcopy:main
-    ableton = cli.ableton:main
-    pretty-collection = cli.pretty_collection:main
-    pretty-history = cli.pretty_history:main
-    pretty-playlists = cli.pretty_playlists:main
+    stemgen = stemgen.cli:main
+    stem = stemgen.stem:main
+    stemsep = stemgen.stemsep:main
+    stemtag = stemgen.stemtag:main
+    stemcopy = stemgen.stemcopy:main
+    ableton = stemgen.ableton:main
+    pretty-collection = stemgen.pretty_collection:main
+    pretty-history = stemgen.pretty_history:main
+    pretty-playlists = stemgen.pretty_playlists:main
