+++ stemgen/__init__.py
@@ -1,0 +1,8 @@
+"""CLI package for Stemgen.""" 
+
+__version__ = "2.0.0"
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
