+++ stemgen/metadata.py
