+++ stemgen/pretty_collection.py
